raw conversation of AssertingOnSetOnlyProperties specs
